+++ lib/src/paint/light_paint_rect.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:tutorial_coach_mark/target_position.dart';
+import 'package:tutorial_coach_mark/src/target/target_position.dart';
 
 class LightPaintRect extends CustomPainter {
   final double progress;
@@ -38,8 +38,7 @@ class LightPaintRect extends CustomPainter {
 
     double h = maxSise * (1 - progress) + target.size.height + offset;
 
-    RRect rrect = RRect.fromRectAndRadius(
-        Rect.fromLTWH(x, y, w, h), Radius.circular(radius));
+    RRect rrect = RRect.fromRectAndRadius(Rect.fromLTWH(x, y, w, h), Radius.circular(radius));
     canvas.drawRRect(rrect, _paintFocus);
     canvas.restore();
   }