+++ lib/src/target/target_content.dart
@@ -1,5 +1,17 @@
 import 'package:flutter/widgets.dart';
-import 'package:tutorial_coach_mark/src/content/custom_content_position.dart';
+
+class CustomTargetContentPosition {
+  CustomTargetContentPosition({
+    this.top,
+    this.left,
+    this.bottom,
+  });
+  final double top, left, bottom;
+  @override
+  String toString() {
+    return 'CustomTargetPosition{top: $top, left: $left, bottom: $bottom}';
+  }
+}
 
 enum ContentAlign { top, bottom, left, right, custom }
 
@@ -11,7 +23,7 @@ class TargetContent {
   }) : assert(child != null && !(align == ContentAlign.custom && customPosition == null));
 
   final ContentAlign align;
-  final CustomContentPosition customPosition;
+  final CustomTargetContentPosition customPosition;
   final Widget child;
   @override
   String toString() {