+++ lib/src/util.dart
@@ -2,6 +2,8 @@ import 'package:flutter/widgets.dart';
 import 'package:tutorial_coach_mark/src/target/target_focus.dart';
 import 'package:tutorial_coach_mark/src/target/target_position.dart';
 
+enum ShapeLightFocus { Circle, RRect }
+
 TargetPosition getTargetCurrent(TargetFocus target) {
   if (target.keyTarget != null) {
     var key = target.keyTarget;