+++ lib/tutorial_coach_mark.dart
@@ -1,11 +1,12 @@
 library tutorial_coach_mark;
 
 import 'package:flutter/material.dart';
-import 'package:tutorial_coach_mark/target_focus.dart';
-import 'package:tutorial_coach_mark/tutorial_coach_mark_widget.dart';
+import 'package:tutorial_coach_mark/src/target/target_focus.dart';
+import 'package:tutorial_coach_mark/src/widgets/tutorial_coach_mark_widget.dart';
 
-export 'package:tutorial_coach_mark/content_target.dart';
-export 'package:tutorial_coach_mark/target_focus.dart';
+export 'package:tutorial_coach_mark/src/target/target_content.dart';
+export 'package:tutorial_coach_mark/src/target/target_focus.dart';
+export 'package:tutorial_coach_mark/src/util.dart';
 
 class TutorialCoachMark {
   final BuildContext _context;
@@ -14,7 +15,7 @@ class TutorialCoachMark {
   final Function(TargetFocus) onClickOverlay;
   final Function() onFinish;
   final double paddingFocus;
-  final Function() onClickSkip;
+  final Function() onSkip;
   final AlignmentGeometry alignSkip;
   final String textSkip;
   final TextStyle textStyleSkip;
@@ -35,7 +36,7 @@ class TutorialCoachMark {
     this.onClickOverlay,
     this.onFinish,
     this.paddingFocus = 10,
-    this.onClickSkip,
+    this.onSkip,
     this.alignSkip = Alignment.bottomRight,
     this.textSkip = "SKIP",
     this.textStyleSkip = const TextStyle(color: Colors.white),
@@ -54,7 +55,7 @@ class TutorialCoachMark {
           clickTarget: onClickTarget,
           clickOverlay: onClickOverlay,
           paddingFocus: paddingFocus,
-          clickSkip: skip,
+          onClickSkip: skip,
           alignSkip: alignSkip,
           textSkip: textSkip,
           textStyleSkip: textStyleSkip,
@@ -70,10 +71,14 @@ class TutorialCoachMark {
   }
 
   void show() {
-    if (_overlayEntry == null) {
-      _overlayEntry = _buildOverlay();
-      Overlay.of(_context).insert(_overlayEntry);
-    }
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      Future.delayed(Duration.zero, () {
+        if (_overlayEntry == null) {
+          _overlayEntry = _buildOverlay();
+          Overlay.of(_context).insert(_overlayEntry);
+        }
+      });
+    });
   }
 
   void finish() {
@@ -82,9 +87,11 @@ class TutorialCoachMark {
   }
 
   void skip() {
-    onClickSkip?.call();
+    onSkip?.call();
     _removeOverlay();
   }
+
+  bool get isShowing => _overlayEntry != null;
 
   void next() => _widgetKey?.currentState?.next();
   void previous() => _widgetKey?.currentState?.previous();